Start avoiding floats in game logic.
The idea is to make the core deterministic, so that replays can be made
by simply recording RNG state at the start and player actions throughout
the game.

This will also enable Cool Things like verified leaderboards without
requiring people to play on a remote server.

NB I don't know if decimal will be up to the task, or whether it'll have
to be replaced by custom fixed point type, but at least it will mark
data that is supposed to be used in the game logic.